fix: removed hook -> inside branchtagselector because just one usecase
